change ubuntu-latest to macos-latest
The ubuntu-latest package gives some ResolvePackageNotFound errors for some packages. [This StackOverflow issue](https://stackoverflow.com/questions/49154899/resolvepackagenotfound-create-env-using-conda-and-yml-file-on-macos) suggests that it could be because of differences in the OS, because the same environment.yml works for Mac (checked with my laptop). 

This commit is a test to see if maybe changing the OS works...