stop using condaenv for ci
Keeping in mind that the CI is constantly failing for building conda envs and the lack of synchronization between pip dependencies and conda dependencies within conda environments, I decided i
t would be easier to remove the conda-based CI, i.e. the workflow yml and the environment yml. I
 will replace this with a requirements.txt and a pypi workflow yml, which I trust to work based
on previous experience. The pypi workflow will be introduced in a future (possibly next in chronology) commit.