display radiative cooling to heating rates for Fig. 3
For the revisions that came in, the 2nd comment (as well as 9th) by Reviewer-1, showed that
there might be some confusion about interpreting the radiative cooling profiles in Fig. 3 as
the convention is to show radiative heating rates. Therefore, the figure is changed accordingly,
along with some changes in the manuscript text itself.

The relevant response to the reviewer's comment is provided here for reference:

"Yes, the reviewer understood correctly that QLW was defined positive for cooling for Fig. 3
and throughout the manuscript. However, considering that the confusion probably arose due
to the convention of using heating rates, we have now switched to showing QLW as heating
rates. This change is now reflected in Fig. 3, in its caption as well as on line 157 which has
been modified to say “net longwave heating rate, Q′LW” "